// packages/business/src/pokemon/external/modules/specie/interface.ts
export interface PokemonSpecieAttributes {
    habitat: string;
    is_baby: boolean;
    shape_url: string;
    shape_name: string;
    is_mythical: boolean;
    gender_rate: number;
    is_legendary: boolean;
    capture_rate: number;
    hatch_counter: number;
    base_happiness: number;
    evolution_chain_url: string;
    evolves_from_species: string;
    has_gender_differences: boolean;
}